requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0

# CLI
rich>=13.7.0
typer>=0.12.0

# Scraping
yt-dlp>=2023.12.30

# Jupyter Notebook
jupyter>=1.0.0
notebook>=7.0.0
pandas>=2.0.0
ipykernel>=6.0.0

# LLM (opcional - pode usar requests direto)
# google-generativeai>=0.3.0  # Para Gemini com SDK oficial
